Update several CHRE sample apps
Adds logging of sensor sample accuracy to sensor world and adds
explicit casts to underlying types when printing a scoped enum to
test nanoapps.

Bug: 171916502
Test: presubmits
Merged-In: I356dc5240bff263ba5679a74eeea8ce75755799a

+++ apps/test/common/chre_settings_test/src/chre_settings_test_manager.cc
@@ -159,7 +159,7 @@ bool Manager::isFeatureSupported(Feature feature) {
       break;
     }
     default:
-      LOGE("Unknown feature %" PRIu8, feature);
+      LOGE("Unknown feature %" PRIu8, static_cast<uint8_t>(feature));
   }
 
   return supported;
@@ -210,7 +210,8 @@ void Manager::handleStartTestMessage(uint16_t hostEndpointId, Feature feature,
     bool success = false;
     if (step == TestStep::SETUP) {
       if (feature != Feature::WIFI_RTT) {
-        LOGE("Unexpected feature %" PRIu8 " for test step", feature);
+        LOGE("Unexpected feature %" PRIu8 " for test step",
+             static_cast<uint8_t>(feature));
       } else {
         success = chreWifiRequestScanAsyncDefault(&kWifiScanningCookie);
       }
@@ -287,14 +288,15 @@ bool Manager::startTestForFeature(Feature feature) {
       break;
     }
     default:
-      LOGE("Unknown feature %" PRIu8, feature);
+      LOGE("Unknown feature %" PRIu8, static_cast<uint8_t>(feature));
       return false;
   }
 
   if (!success) {
-    LOGE("Failed to make request for test feature %" PRIu8, feature);
+    LOGE("Failed to make request for test feature %" PRIu8,
+         static_cast<uint8_t>(feature));
   } else {
-    LOGI("Starting test for feature %" PRIu8, feature);
+    LOGI("Starting test for feature %" PRIu8, static_cast<uint8_t>(feature));
   }
 
   return success;
@@ -334,7 +336,7 @@ void Manager::handleWifiAsyncResult(const chreAsyncResult *result) {
       }
       if (mTestSession->feature != Feature::WIFI_SCANNING) {
         LOGE("Unexpected WiFi scan async result: test feature %" PRIu8,
-             mTestSession->feature);
+             static_cast<uint8_t>(mTestSession->feature));
       } else {
         success = validateAsyncResult(
             result, static_cast<const void *>(&kWifiScanningCookie));
@@ -344,7 +346,7 @@ void Manager::handleWifiAsyncResult(const chreAsyncResult *result) {
     case CHRE_WIFI_REQUEST_TYPE_RANGING: {
       if (mTestSession->feature != Feature::WIFI_RTT) {
         LOGE("Unexpected WiFi ranging async result: test feature %" PRIu8,
-             mTestSession->feature);
+             static_cast<uint8_t>(mTestSession->feature));
       } else {
         success = validateAsyncResult(
             result, static_cast<const void *>(&kWifiRttCookie));
@@ -394,7 +396,7 @@ void Manager::handleGnssAsyncResult(const chreAsyncResult *result) {
     case CHRE_GNSS_REQUEST_TYPE_LOCATION_SESSION_START: {
       if (mTestSession->feature != Feature::GNSS_LOCATION) {
         LOGE("Unexpected GNSS location async result: test feature %" PRIu8,
-             mTestSession->feature);
+             static_cast<uint8_t>(mTestSession->feature));
       } else {
         success = validateAsyncResult(
             result, static_cast<const void *>(&kGnssLocationCookie));
@@ -405,7 +407,7 @@ void Manager::handleGnssAsyncResult(const chreAsyncResult *result) {
     case CHRE_GNSS_REQUEST_TYPE_MEASUREMENT_SESSION_START: {
       if (mTestSession->feature != Feature::GNSS_MEASUREMENT) {
         LOGE("Unexpected GNSS measurement async result: test feature %" PRIu8,
-             mTestSession->feature);
+             static_cast<uint8_t>(mTestSession->feature));
       } else {
         success = validateAsyncResult(
             result, static_cast<const void *>(&kGnssMeasurementCookie));
@@ -426,7 +428,7 @@ void Manager::handleWwanCellInfoResult(const chreWwanCellInfoResult *result) {
   // CHRE API requirements.
   if (mTestSession->feature != Feature::WWAN_CELL_INFO) {
     LOGE("Unexpected WWAN cell info result: test feature %" PRIu8,
-         mTestSession->feature);
+         static_cast<uint8_t>(mTestSession->feature));
   } else if (result->cookie != &kWwanCellInfoCookie) {
     LOGE("Unexpected cookie on WWAN cell info result");
   } else if (result->errorCode != CHRE_ERROR_NONE) {
